reconsidering the whole welder component thing

+++ code/obj.dm
@@ -346,10 +346,19 @@
 				user.u_equip(T)
 				qdel(T)
 			return
-		if (SEND_SIGNAL(src, COMSIG_ITEM_ATTACK_OBJECT, C, user, 1, 1) & ITEM_EFFECT_WELD)
-			boutput(user, "<span class='notice'>Slicing lattice joints ...</span>")
-			new /obj/item/rods/steel(src.loc)
-			qdel(src)
+		var/list/burn_return = list(HAS_EFFECT = ITEM_EFFECT_NOTHING, EFFECT_RESULT = ITEM_EFFECT_FAILURE)
+		SEND_SIGNAL(this = W, COMSIG_ITEM_ATTACK_OBJECT, src, user = user, results = burn_return, use_amt = 1, noisy = 1)
+		if(burn_return[HAS_EFFECT] & ITEM_EFFECT_WELD)
+			if(burn_return[EFFECT_RESULT] & ITEM_EFFECT_NO_FUEL)
+				boutput(user, "<span class='notice'>\the [W] is out of fuel!</span>")
+			else if(burn_return[EFFECT_RESULT] & ITEM_EFFECT_NOT_ENOUGH_FUEL)
+				boutput(user, "<span class='notice'>\the [W] doesn't have enough fuel!</span>")
+			else if(burn_return[EFFECT_RESULT] & ITEM_EFFECT_NOT_ON)
+				boutput(user, "<span class='notice'>\the [W] isn't lit!</span>")
+			else
+				boutput(user, "<span class='notice'>Slicing lattice joints ...</span>")
+				new /obj/item/rods/steel(src.loc)
+				qdel(src)
 		if (istype(C, /obj/item/rods))
 			var/obj/item/rods/R = C
 			if (R.amount >= 2)
@@ -379,12 +388,21 @@
 			return
 
 	attackby(obj/item/W as obj, mob/user as mob)
-		if (SEND_SIGNAL(src, COMSIG_ITEM_ATTACK_OBJECT, W, user, 1, 1) & ITEM_EFFECT_WELD)
-			boutput(user, "<span class='notice'>You disassemble the barricade.</span>")
-			var/obj/item/rods/R = new /obj/item/rods/steel(src.loc)
-			R.amount = src.strength
-			qdel(src)
-			return
+		var/list/burn_return = list(HAS_EFFECT = ITEM_EFFECT_NOTHING, EFFECT_RESULT = ITEM_EFFECT_FAILURE)
+		SEND_SIGNAL(this = W, COMSIG_ITEM_ATTACK_OBJECT, src, user = user, results = burn_return, use_amt = 1, noisy = 1)
+		if(burn_return[HAS_EFFECT] & ITEM_EFFECT_WELD)
+			if(burn_return[EFFECT_RESULT] & ITEM_EFFECT_NO_FUEL)
+				boutput(user, "<span class='notice'>\the [W] is out of fuel!</span>")
+			else if(burn_return[EFFECT_RESULT] & ITEM_EFFECT_NOT_ENOUGH_FUEL)
+				boutput(user, "<span class='notice'>\the [W] doesn't have enough fuel!</span>")
+			else if(burn_return[EFFECT_RESULT] & ITEM_EFFECT_NOT_ON)
+				boutput(user, "<span class='notice'>\the [W] isn't lit!</span>")
+			else
+				boutput(user, "<span class='notice'>You disassemble the barricade.</span>")
+				var/obj/item/rods/R = new /obj/item/rods/steel(src.loc)
+				R.amount = src.strength
+				qdel(src)
+				return
 		else if (istype(W,/obj/item/rods))
 			var/obj/item/rods/R = W
 			var/difference = 5 - src.strength
